requirements.txt
openai>=1.0.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
pandas>=2.0.0
requests>=2.31.0
Pillow>=10.0.0
tenacity>=8.2.0
tqdm>=4.66.0
